+++ tilepicker.h
@@ -2,8 +2,8 @@
 #define TILEPICKER_H
 
 void tilepicker_recalc (void);
-bool tilepicker_first (int *x, int *y, int *wd, int *ht, int *zoom, float p[4][3]);
-bool tilepicker_next (int *x, int *y, int *wd, int *ht, int *zoom, float p[4][3]);
+bool tilepicker_first (float *x, float *y, float *wd, float *ht, int *zoom, float p[4][3]);
+bool tilepicker_next (float *x, float *y, float *wd, float *ht, int *zoom, float p[4][3]);
 void tilepicker_destroy (void);
 
 #endif	/* TILEPICKER_H */
